+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add Return overload for simpler SpireReturn from void methods
 * Fix ClassPool not returning modified versions of classes after patching is complete
 * Add ModTheSpire application icon
+* Alter Slay the Spire window title
 
 #### v3.19.1 ####
 * Fix crash if mod list file is corrupted